use float for resolution numbes

+++ src/hf_hydrodata/data_access.py
@@ -1126,7 +1126,7 @@ def grid_to_latlng(grid: str, *args) -> List[float]:
     grid_row = get_table_row("grid", id=grid.lower())
     if grid_row is None:
         raise ValueError(f"No such grid {grid} available.")
-    grid_resolution = int(grid_row["resolution_meters"])
+    grid_resolution = float(grid_row["resolution_meters"])
     if len(args) == 0:
         raise ValueError("At least two x, y values must be provided.")
     if len(args) % 2 == 1:
@@ -1166,7 +1166,7 @@ def latlng_to_grid(grid: str, *args) -> List[float]:
     grid_row = get_table_row("grid", id=grid.lower())
     if grid_row is None:
         raise ValueError(f"No such grid {grid} available.")
-    grid_resolution = int(grid_row["resolution_meters"])
+    grid_resolution = float(grid_row["resolution_meters"])
     if len(args) == 0:
         raise ValueError("At least two x, y values must be provided.")
     if len(args) % 2 == 1:
